+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Fixed Supervise mode
 * Separate stderr and stdout in dispatcher
 * Fix history server stderr/stdout. Now is possible to set log level through SPARK_LOG_LEVEL
+* Secret folder path configurable
 
 ## 2.2.0.4 (January 11, 2018)